Moved the Systen files to entity component system
